+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/_format.py
@@ -11,6 +11,11 @@ def aks_list_table_format(results):
     return [_aks_table_format(r) for r in results]
 
 
+def osa_list_table_format(results):
+    """"Format a list of OpenShift managed clusters as summary results for display with "-o table"."""
+    return [_osa_table_format(r) for r in results]
+
+
 def aks_show_table_format(result):
     """Format a managed cluster as summary results for display with "-o table"."""
     return [_aks_table_format(result)]
@@ -24,6 +29,21 @@ def _aks_table_format(result):
         location: location,
         resourceGroup: resourceGroup,
         kubernetesVersion: kubernetesVersion,
+        provisioningState: provisioningState,
+        fqdn: fqdn
+    }""")
+    # use ordered dicts so headers are predictable
+    return parsed.search(result, Options(dict_cls=OrderedDict))
+
+
+def _osa_table_format(result):
+    from jmespath import compile as compile_jmes, Options
+
+    parsed = compile_jmes("""{
+        name: name,
+        location: location,
+        resourceGroup: resourceGroup,
+        openShiftVersion: openShiftVersion,
         provisioningState: provisioningState,
         fqdn: fqdn
     }""")