+++ src/command_modules/azure-cli-iot/azure/cli/command_modules/iot/_help.py
@@ -155,7 +155,7 @@ helps['iot dps linked-hub update'] = """
     examples:
         - name: Update linked IoT hub 'MyLinkedHub.azure-devices.net' in an Azure IoT Hub device provisioning service
           text: >
-            az iot dps linked-hub update --dps-name MyDps --resource-group MyResourceGroup --name MyLinkedHub.azure-devices.net
+            az iot dps linked-hub update --dps-name MyDps --resource-group MyResourceGroup --linked-hub MyLinkedHub.azure-devices.net
             --allocation-weight 10 --apply-allocation-policy True
 """
 
@@ -174,7 +174,7 @@ helps['iot dps linked-hub show'] = """
     examples:
         - name: Show details of linked IoT hub 'MyLinkedHub' in an Azure IoT Hub device provisioning service
           text: >
-            az iot dps linked-hub show --dps-name MyDps --resource-group MyResourceGroup --name MyLinkedHub
+            az iot dps linked-hub show --dps-name MyDps --resource-group MyResourceGroup --linked-hub MyLinkedHub
 """
 
 helps['iot dps linked-hub delete'] = """
@@ -183,7 +183,7 @@ helps['iot dps linked-hub delete'] = """
     examples:
         - name: Delete linked IoT hub 'MyLinkedHub' in an Azure IoT Hub device provisioning service
           text: >
-            az iot dps linked-hub delete --dps-name MyDps --resource-group MyResourceGroup --name MyLinkedHub
+            az iot dps linked-hub delete --dps-name MyDps --resource-group MyResourceGroup --linked-hub MyLinkedHub
 """
 
 helps['iot dps certificate'] = """
@@ -528,14 +528,14 @@ helps['iot hub routing-endpoint create'] = """
         - name: Add a new endpoint "E2" of type EventHub to "MyIotHub" IoT Hub.
           text: >
             az iot hub routing-endpoint create --resource-group MyResourceGroup --hub-name MyIotHub
-            --endpoint-name E2 --endpoint-type eventhub --endpoint-resource-group [Resource Group]
-            --endpoint-subscription-id [SubscriptionId] --connection-string [Connection String]
+            --endpoint-name E2 --endpoint-type eventhub --endpoint-resource-group {ResourceGroup}
+            --endpoint-subscription-id {SubscriptionId} --connection-string {ConnectionString}
         - name: Add a new endpoint "S1" of type AzureStorageContainer to "MyIotHub" IoT Hub.
-          text: >
-            az iot hub routing-endpoint create --resource-group MyResourceGroup --hub-name MyIotHub
-            --endpoint-name S1 --endpoint-type azurestoragecontainer --endpoint-resource-group [Resource Group]
-            --endpoint-subscription-id [SubscriptionId] --connection-string [Connection String]
-            --container-name [Container Name]
+          text: |
+            az iot hub routing-endpoint create --resource-group MyResourceGroup --hub-name MyIotHub \\
+            --endpoint-name S1 --endpoint-type azurestoragecontainer --endpoint-resource-group "[Resource Group]" \\
+            --endpoint-subscription-id {SubscriptionId} --connection-string {ConnectionString} \\
+            --container-name {ContainerName}
 """
 
 helps['iot hub routing-endpoint list'] = """
@@ -559,9 +559,9 @@ helps['iot hub routing-endpoint show'] = """
     long-summary: Get information on a specific endpoint in your IoT Hub
     examples:
         - name: Get an endpoint information from "MyIotHub" IoT Hub.
-          text: >
-            az iot hub routing-endpoint show --resource-group MyResourceGroup --hub-name MyIotHub
-            --endpoint-name [Endpoint Name]
+          text: |
+            az iot hub routing-endpoint show --resource-group MyResourceGroup --hub-name MyIotHub \\
+            --endpoint-name {endpointName}
 """
 
 helps['iot hub routing-endpoint delete'] = """
@@ -624,7 +624,7 @@ helps['iot hub route show'] = """
     examples:
         - name: Get an route information from "MyIotHub" IoT Hub.
           text: >
-            az iot hub route show -g MyResourceGroup --hub-name MyIotHub --route-name [Route Name]
+            az iot hub route show -g MyResourceGroup --hub-name MyIotHub --route-name {routeName}
 """
 
 helps['iot hub route delete'] = """
@@ -666,4 +666,19 @@ helps['iot hub route update'] = """
           text: >
             az iot hub route update -g MyResourceGroup --hub-name MyIotHub
             --source-type DeviceMessages --route-name R1
+"""
+
+helps['iot hub devicestream'] = """
+    type: group
+    short-summary: Manage device streams of an IoT hub.
+"""
+
+helps['iot hub devicestream show'] = """
+    type: command
+    short-summary: Get IoT Hub's device streams endpoints.
+    long-summary: Get IoT Hub's device streams endpoints.
+    examples:
+        - name: Get all the device streams from "MyIotHub" IoT Hub.
+          text: >
+            az iot hub devicestream show -n MyIotHub
 """