+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/tests/latest/test_network_commands.py
@@ -2108,6 +2108,48 @@ class NetworkTrafficManagerScenarioTest(ScenarioTest):
 
         self.cmd('network traffic-manager profile delete -g {rg} -n {tm}')
 
+    @ResourceGroupPreparer('cli_test_traffic_manager_subnet')
+    def test_network_traffic_manager_subnet_routing(self, resource_group):
+
+        self.kwargs.update({
+            'tm': 'tm1',
+            'endpoint': 'ep1',
+            'dns': self.create_random_name('testtm', 20),
+            'pip': 'ip1',
+            'ip_dns': self.create_random_name('testpip', 20)
+        })
+
+        self.cmd('network traffic-manager profile create -n {tm} -g {rg} --routing-method subnet --unique-dns-name {dns} --custom-headers foo=bar --status-code-ranges 200-202', checks=[
+            self.check('TrafficManagerProfile.monitorConfig.expectedStatusCodeRanges[0].min', 200),
+            self.check('TrafficManagerProfile.monitorConfig.expectedStatusCodeRanges[0].max', 202),
+            self.check('TrafficManagerProfile.monitorConfig.customHeaders[0].name', 'foo'),
+            self.check('TrafficManagerProfile.monitorConfig.customHeaders[0].value', 'bar')
+        ])
+        self.kwargs['ip_id'] = self.cmd('network public-ip create -g {rg} -n {pip} --dns-name {ip_dns} --query publicIp.id').get_output_in_json()
+        self.cmd('network traffic-manager profile update -n {tm} -g {rg} --status-code-ranges 200-204 --custom-headers foo=doo test=best', checks=[
+            self.check('monitorConfig.expectedStatusCodeRanges[0].min', 200),
+            self.check('monitorConfig.expectedStatusCodeRanges[0].max', 204),
+            self.check('monitorConfig.customHeaders[0].name', 'foo'),
+            self.check('monitorConfig.customHeaders[0].value', 'doo'),
+            self.check('monitorConfig.customHeaders[1].name', 'test'),
+            self.check('monitorConfig.customHeaders[1].value', 'best')
+        ])
+
+        # Endpoint tests
+        self.cmd('network traffic-manager endpoint create -n {endpoint} --profile-name {tm} -g {rg} --type azureEndpoints --target-resource-id {ip_id} --subnets 10.0.0.0 --custom-headers test=best', checks=[
+            self.check('customHeaders[0].name', 'test'),
+            self.check('customHeaders[0].value', 'best'),
+            self.check('subnets[0].first', '10.0.0.0')
+        ])
+        self.cmd('network traffic-manager endpoint update -n {endpoint} --type azureEndpoints --profile-name {tm} -g {rg} --subnets 10.0.0.0:24', checks=[
+            self.check('subnets[0].first', '10.0.0.0'),
+            self.check('subnets[0].scope', '24')
+        ])
+        self.cmd('network traffic-manager endpoint update -n {endpoint} --type azureEndpoints --profile-name {tm} -g {rg} --subnets 10.0.0.0-11.0.0.0', checks=[
+            self.check('subnets[0].first', '10.0.0.0'),
+            self.check('subnets[0].last', '11.0.0.0')
+        ])
+
 
 class NetworkWatcherConfigureScenarioTest(LiveScenarioTest):
 
@@ -2125,11 +2167,6 @@ class NetworkWatcherScenarioTest(ScenarioTest):
     def _mock_thread_count():
         return 1
 
-    def _configure_network_watcher(self):
-        # ensure network watcher RG exists and is configured for our location
-        self.cmd('group create -g NetworkWatcherRg -l westcentralus')
-        self.cmd('network watcher configure -g NetworkWatcherRg --locations westcentralus --enabled')
-
     @mock.patch('azure.cli.command_modules.vm._actions._get_thread_count', _mock_thread_count)
     @ResourceGroupPreparer(name_prefix='cli_test_nw_vm', location='westcentralus')
     @AllowLargeResponse()
@@ -2142,7 +2179,7 @@ class NetworkWatcherScenarioTest(ScenarioTest):
             'capture': 'capture1',
             'private-ip': '10.0.0.9'
         })
-        self._configure_network_watcher()
+
         vm = self.cmd('vm create -g {rg} -n {vm} --image UbuntuLTS --authentication-type password --admin-username deploy --admin-password PassPass10!) --nsg {nsg} --private-ip-address {private-ip}').get_output_in_json()
         self.kwargs['vm_id'] = vm['id']
         self.cmd('vm extension set -g {rg} --vm-name {vm} -n NetworkWatcherAgentLinux --publisher Microsoft.Azure.NetworkWatcher')
@@ -2165,7 +2202,6 @@ class NetworkWatcherScenarioTest(ScenarioTest):
             'sa': storage_account
         })
 
-        self._configure_network_watcher()
         self.cmd('network nsg create -g {rg} -n {nsg}')
         self.cmd('network watcher flow-log configure -g {rg} --nsg {nsg} --enabled --retention 5 --storage-account {sa}')
         self.cmd('network watcher flow-log configure -g {rg} --nsg {nsg} --retention 0')
@@ -2181,7 +2217,7 @@ class NetworkWatcherScenarioTest(ScenarioTest):
             'vm': 'vm1',
             'capture': 'capture1'
         })
-        self._configure_network_watcher()
+
         self.cmd('vm create -g {rg} -n {vm} --image UbuntuLTS --authentication-type password --admin-username deploy --admin-password PassPass10!) --nsg {vm}')
         self.cmd('vm extension set -g {rg} --vm-name {vm} -n NetworkWatcherAgentLinux --publisher Microsoft.Azure.NetworkWatcher')
 
@@ -2202,7 +2238,6 @@ class NetworkWatcherScenarioTest(ScenarioTest):
             'loc': resource_group_location,
             'sa': storage_account
         })
-        self._configure_network_watcher()
 
         # set up resource to troubleshoot
         self.cmd('storage container create -n troubleshooting --account-name {sa}')
@@ -2227,7 +2262,6 @@ class NetworkWatcherScenarioTest(ScenarioTest):
             'vm3': 'vm3',
             'cm': 'cm1'
         })
-        self._configure_network_watcher()
 
         self.cmd('vm create -g {rg} -n {vm2} --image UbuntuLTS --authentication-type password --admin-username deploy --admin-password PassPass10!) --nsg {vm2}')
         self.cmd('vm extension set -g {rg} --vm-name {vm2} -n NetworkWatcherAgentLinux --publisher Microsoft.Azure.NetworkWatcher')