+++ src/command_modules/azure-cli-resource/azure/cli/command_modules/resource/_help.py
@@ -326,62 +326,62 @@ helps['policy definition create'] = """
             examples:
                 - name: Create a read-only policy.
                   text: |
-                    az policy definition create --name readOnlyStorage --rules '{
-                        "if":
-                        {
-                            "field": "type",
-                            "equals": "Microsoft.Storage/storageAccounts/write"
-                        },
-                        "then":
-                        {
-                            "effect": "deny"
-                        }
+                    az policy definition create --name readOnlyStorage --rules '{ \\
+                        "if": \\
+                        { \\
+                            "field": "type", \\
+                            "equals": "Microsoft.Storage/storageAccounts/write" \\
+                        }, \\
+                        "then": \\
+                        { \\
+                            "effect": "deny" \\
+                        } \\
                     }'
                 - name: Create a policy parameter definition.
                   text: |
-                    az policy definition create --name allowedLocations --rules '{
-                        "if": {
-                            "allOf": [
-                                {
-                                    "field": "location",
-                                    "notIn": "[parameters('listOfAllowedLocations')]"
-                                },
-                                {
-                                    "field": "location",
-                                    "notEquals": "global"
-                                },
-                                {
-                                    "field": "type",
-                                    "notEquals": "Microsoft.AzureActiveDirectory/b2cDirectories"
-                                }
-                            ]
-                        },
-                        "then": {
-                            "effect": "deny"
-                        }
+                    az policy definition create --name allowedLocations --rules '{ \\
+                        "if": { \\
+                            "allOf": [ \\
+                                { \\
+                                    "field": "location", \\
+                                    "notIn": "[parameters('listOfAllowedLocations')]" \\
+                                }, \\
+                                { \\
+                                    "field": "location", \\
+                                    "notEquals": "global" \\
+                                }, \\
+                                { \\
+                                    "field": "type", \\
+                                    "notEquals": "Microsoft.AzureActiveDirectory/b2cDirectories" \\
+                                } \\
+                            ] \\
+                        }, \\
+                        "then": { \\
+                            "effect": "deny" \\
+                        } \\
                     }' \\
-                    --params '{
-                        "allowedLocations": {
-                            "type": "array",
-                            "metadata": {
-                                "description": "The list of locations that can be specified when deploying resources",
-                                "strongType": "location",
-                                "displayName": "Allowed locations"
-                            }
-                        }
+                    --params '{ \\
+                        "allowedLocations": { \\
+                            "type": "array", \\
+                            "metadata": { \\
+                                "description": "The list of locations that can be specified when deploying resources", \\
+                                "strongType": "location", \\
+                                "displayName": "Allowed locations" \\
+                            } \\
+                        } \\
                     }'
                 - name: Create a read-only policy that can be applied within a management group.
                   text: |
-                    az policy definition create -n readOnlyStorage --management-group 'MyManagementGroup' --rules '{
-                        "if":
-                        {
-                            "field": "type",
-                            "equals": "Microsoft.Storage/storageAccounts/write"
-                        },
-                        "then":
-                        {
-                            "effect": "deny"
-                        }
+                    az policy definition create -n readOnlyStorage --management-group 'MyManagementGroup' --rules '{ \\
+                        "if": \\
+                        { \\
+                            "field": "type", \\
+                            "equals": "Microsoft.Storage/storageAccounts/write" \\
+                        }, \\
+                        "then": \\
+                        { \\
+                            "effect": "deny" \\
+                        } \\
                     }'
 """
 helps['policy definition delete'] = """
@@ -420,17 +420,17 @@ helps['policy set-definition create'] = """
             examples:
                 - name: Create a policy set definition.
                   text: |
-                    az policy set-definition create -n readOnlyStorage --definitions '[
-                            {
-                                "policyDefinitionId": "/subscriptions/mySubId/providers/Microsoft.Authorization/policyDefinitions/storagePolicy"
-                            }
+                    az policy set-definition create -n readOnlyStorage --definitions '[ \\
+                            { \\
+                                "policyDefinitionId": "/subscriptions/mySubId/providers/Microsoft.Authorization/policyDefinitions/storagePolicy" \\
+                            } \\
                         ]'
                 - name: Create a policy set definition to be used by a subscription.
                   text: |
-                    az policy set-definition create -n readOnlyStorage --subscription '0b1f6471-1bf0-4dda-aec3-111122223333' --definitions '[
-                            {
-                                "policyDefinitionId": "/subscriptions/mySubId/providers/Microsoft.Authorization/policyDefinitions/storagePolicy"
-                            }
+                    az policy set-definition create -n readOnlyStorage --subscription '0b1f6471-1bf0-4dda-aec3-111122223333' --definitions '[ \\
+                            { \\
+                                "policyDefinitionId": "/subscriptions/mySubId/providers/Microsoft.Authorization/policyDefinitions/storagePolicy" \\
+                            } \\
                         ]'
 """
 helps['policy set-definition delete'] = """
@@ -468,25 +468,25 @@ helps['policy assignment create'] = """
               subscription:      /subscriptions/0b1f6471-1bf0-4dda-aec3-111122223333
               resource group:    /subscriptions/0b1f6471-1bf0-4dda-aec3-111122223333/resourceGroups/myGroup
               resource:          /subscriptions/0b1f6471-1bf0-4dda-aec3-111122223333/resourceGroups/myGroup/providers/Microsoft.Compute/virtualMachines/myVM
-                az policy assignment create --scope '/providers/Microsoft.Management/managementGroups/MyManagementGroup' --policy {PolicyName} -p '{
-                    "allowedLocations": {
-                        "value": [
-                            "australiaeast",
-                            "eastus",
-                            "japaneast"
-                        ]
-                    }
+                az policy assignment create --scope '/providers/Microsoft.Management/managementGroups/MyManagementGroup' --policy {PolicyName} -p '{ \\
+                    "allowedLocations": { \\
+                        "value": [ \\
+                            "australiaeast", \\
+                            "eastus", \\
+                            "japaneast" \\
+                        ] \\
+                    } \\
                 }'
         - name: Create a resource policy assignment and provide rule parameter values.
           text: |
-                az policy assignment create --policy {PolicyName} -p '{
-                    "allowedLocations": {
-                        "value": [
-                            "australiaeast",
-                            "eastus",
-                            "japaneast"
-                        ]
-                    }
+                az policy assignment create --policy {PolicyName} -p '{ \\
+                    "allowedLocations": { \\
+                        "value": [ \\
+                            "australiaeast", \\
+                            "eastus", \\
+                            "japaneast" \\
+                        ] \\
+                    } \\
                 }'
 """
 helps['policy assignment delete'] = """
@@ -580,20 +580,20 @@ helps['resource create'] = """
     examples:
        - name: Create an API app by providing a full JSON configuration.
          text: |
-            az resource create -g myRG -n myApiApp --resource-type Microsoft.web/sites --is-full-object --properties '{
-                        "kind": "api",
-                        "location": "West US",
-                        "properties": {
-                            "serverFarmId": "/subscriptions/{SubID}/resourcegroups/{ResourceGroup}/providers/Microsoft.Web/serverfarms/{ServicePlan}"
-                        }
+            az resource create -g myRG -n myApiApp --resource-type Microsoft.web/sites --is-full-object --properties '{ \\
+                        "kind": "api", \\
+                        "location": "West US", \\
+                        "properties": { \\
+                            "serverFarmId": "/subscriptions/{SubID}/resourcegroups/{ResourceGroup}/providers/Microsoft.Web/serverfarms/{ServicePlan}" \\
+                        } \\
                     }'
        - name: Create a resource by loading JSON configuration from a file.
          text: >
             az resource create -g myRG -n myApiApp --resource-type Microsoft.web/sites --is-full-object --properties @jsonConfigFile
        - name: Create a web app with the minimum required configuration information.
          text: |
-            az resource create -g myRG -n myWeb --resource-type Microsoft.web/sites --properties '{
-                    "serverFarmId":"/subscriptions/{SubID}/resourcegroups/{ResourceGroup}/providers/Microsoft.Web/serverfarms/{ServicePlan}"
+            az resource create -g myRG -n myWeb --resource-type Microsoft.web/sites --properties '{ \\
+                    "serverFarmId":"/subscriptions/{SubID}/resourcegroups/{ResourceGroup}/providers/Microsoft.Web/serverfarms/{ServicePlan}" \\
                 }'
 """
 
@@ -623,10 +623,10 @@ helps['resource invoke-action'] = """
          text: >
             az resource invoke-action --action capture \\
               --ids /subscriptions/{SubID}/resourceGroups/{ResourceGroup}/providers/Microsoft.Compute/virtualMachines/{VMName} \\
-              --request-body '{
-                "vhdPrefix": "myPrefix",
-                "destinationContainerName": "myContainer",
-                "overwriteVhds": true
+              --request-body '{ \\
+                "vhdPrefix": "myPrefix", \\
+                "destinationContainerName": "myContainer", \\
+                "overwriteVhds": true \\
             }'
 """
 
@@ -713,10 +713,10 @@ helps['group deployment create'] = """
             az group deployment create -g MyResourceGroup --template-uri https://myresource/azuredeploy.json --parameters @myparameters.json
         - name: Create a deployment from a local template file, using parameters from a JSON string.
           text: |
-            az group deployment create -g MyResourceGroup --template-file azuredeploy.json --parameters '{
-                    "location": {
-                        "value": "westus"
-                    }
+            az group deployment create -g MyResourceGroup --template-file azuredeploy.json --parameters '{ \\
+                    "location": { \\
+                        "value": "westus" \\
+                    } \\
                 }'
         - name: Create a deployment from a local template, using a parameter file and selectively overriding key/value pairs.
           text: >
@@ -764,10 +764,10 @@ helps['deployment create'] = """
             az deployment create --location WestUS --template-uri https://myresource/azuredeploy.json --parameters @myparameters.json
         - name: Create a deployment from a local template file, using parameters from a JSON string.
           text: |
-            az deployment create --location WestUS --template-file azuredeploy.json --parameters '{
-                    "policyName": {
-                        "value": "policy2"
-                    }
+            az deployment create --location WestUS --template-file azuredeploy.json --parameters '{ \\
+                    "policyName": { \\
+                        "value": "policy2" \\
+                    } \\
                 }'
         - name: Create a deployment from a local template, using a parameter file and selectively overriding key/value pairs.
           text: >