+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -2,10 +2,18 @@
 
 Release History
 ===============
+2.3.1
+++++++
+* `ddos-protection update`: Fix issue with set comprehension causing stack trace.
+
+2.3.0
++++++
+* `traffic-manager profile create/update`: Add support for `--custom-headers` and `--status-code-ranges`. Add support for new routing types: Subnet and Multivalue.
+* `traffic-manager endpoint create/update`: Add support for `--custom-headers` and `--subnets`.
+* `ddos-protection update`: Fix issue where supplying `--vnets ""` to remove vnets caused a strack trace.
 
 2.2.11
 ++++++
-* `ddos-protection update`: Fix issue where supplying `--vnets ""` to remove vnets caused a strack trace.
 * `watcher flow-log configure`: Add support for `--format` and `--log-version`.
 * `dns zone update`: Finished issue where using "" to clear resolution and registration VNets didn't work.